Rename aesara.sparse.opt to aesara.sparse.rewriting
